finitions Section + Project

+++ src/main/java/com/projet/korector/controller/SectionController.java
@@ -1,11 +1,18 @@
 package com.projet.korector.controller;
+import com.projet.korector.entity.ERole;
+import com.projet.korector.entity.Role;
 import com.projet.korector.entity.Section;
+import com.projet.korector.entity.User;
+import com.projet.korector.repository.UserRepository;
 import com.projet.korector.services.SectionService;
+import com.projet.korector.services.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -17,6 +24,8 @@ public class SectionController {
 
     @Autowired
     private SectionService service;
+    @Autowired
+    private UserService userService;
 
     @PostMapping(value = "/createSection" )
     public String createSection(@RequestBody Section section) {
@@ -39,7 +48,14 @@ public class SectionController {
     @DeleteMapping(value = "/deleteSection/{sectionId}")
     public void deleteSection(@PathVariable Long sectionId)
     {
-        service.deleteSection(sectionId);
+        List<User> users =  userService.findAllUser();
+        if(users.isEmpty()){
+            service.deleteSection(sectionId);
+        }
+        else{
+            System.out.println("Impossible de supprimer la section, des utilisateurs y sont rattachés");
+        }
+
     }
 
     @RequestMapping(value = "/SectionByName/{name}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
@@ -54,6 +70,45 @@ public class SectionController {
         return service.getSectionById(id);
     }
 
+    @RequestMapping(value = "/Section/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    public List<User> getStudents(@PathVariable Long id){
+        System.out.println("dans la methode ");
+        List<User> users =  userService.findAllUser();
+        List<User> result = new ArrayList<User>();
+        for (User u:users) {
+            for (Role role : u.getRoles()) {
+                if (role.getName().toString() == ("ROLE_ETUDIANT")) {
+                    for (Section s : u.getSections()) {
+                        if (s.getId() == id) {
+
+                            result.add(u);
+                        }
+                    }
+                }
+            }
+        }
+        return result;
+    }
+
+    @RequestMapping(value = "/SectionTeachers/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
+    public List<User> getTeachers(@PathVariable Long id){
+        System.out.println("dans la methode ");
+        List<User> users =  userService.findAllUser();
+        List<User> result = new ArrayList<User>();
+        for (User u:users) {
+            for (Role role : u.getRoles()) {
+                if (role.getName().toString() == ("ROLE_ENSEIGNANT")) {
+                    for (Section s : u.getSections()) {
+                        if (s.getId() == id) {
+
+                            result.add(u);
+                        }
+                    }
+                }
+            }
+        }
+        return result;
+    }
 
 
 }
